Fix compiler warnings.
* src/preproc/html/pre-html.cpp (alterDeviceTo): Avoid ambiguous
if-else clause.

* src/preproc/grn/main.cpp (sccsid): Comment out.

* src/roff/troff/number.cpp (parse_expr) <case '*'>: Add cast.

* src/devices/xditview/Makefile.sub (EXTRA_CFLAGS): Add `-Dlint' so
that unused static ID arrays don't cause a warning.

(Some) problems reported by Bjarni Ingi Gislason
<bjarniig@rhi.hi.is>.

+++ src/roff/troff/number.cpp
@@ -1,5 +1,5 @@
 // -*- C++ -*-
-/* Copyright (C) 1989, 1990, 1991, 1992, 2001, 2002, 2004, 2009
+/* Copyright (C) 1989-1992, 2001, 2002, 2004, 2009, 2012
    Free Software Foundation, Inc.
      Written by James Clark (jjc@jclark.com)
 
@@ -355,7 +355,7 @@ static int parse_expr(units *v, int scale_indicator,
     case '*':
       if (v2 < 0) {
 	if (*v > 0) {
-	  if (*v > -(unsigned)INT_MIN / -(unsigned)v2)
+	  if ((unsigned)*v > -(unsigned)INT_MIN / -(unsigned)v2)
 	    overflow = 1;
 	}
 	else if (-(unsigned)*v > INT_MAX / -(unsigned)v2)
